pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docquery-backend"
version = "0.1.0"
description = "Backend for DocQuery RAG application"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "pydantic-settings",
    "python-dotenv",
    # Allow modern versions
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-pinecone>=0.2.0",
    "pinecone-client>=5.0.0",
    "sentence-transformers",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
